[iOS] fix: Priority PopoverViewModelType 부분 파일 분리
- Priority에서 UIKit이 필요한 부분만 별도로 분리했습니다
