+++ iOS/HalgoraeDO/Sources/Models/Priority.swift
@@ -5,7 +5,7 @@
 //  Created by woong on 2020/11/26.
 //
 
-import UIKit
+import Foundation
 
 enum Priority: Int, CaseIterable {
     case one = 1
@@ -19,35 +19,6 @@ enum Priority: Int, CaseIterable {
             case .two: return "우선순위 2"
             case .three: return "우선순위 3"
             case .four: return "우선순위 4"
-        }
-    }
-    
-    var color: UIColor {
-        switch self {
-            case .one: return .red
-            case .two: return .blue
-            case .three: return .orange
-            case .four: return .black
-        }
-    }
-}
-
-// MARK: - For PopoverViewModel
-
-extension Priority {
-    struct ViewModel: PopoverViewModelType {
-        var title: String
-        var tintColor: UIColor?
-        var image: UIImage?
-    }
-    
-    var viewModel: ViewModel {
-        let image = UIImage(systemName: "flag.fill")?.scaled(to: .init(width: 30, height: 30))
-        switch self {
-            case .one: return ViewModel(title: title, tintColor: color, image: image)
-            case .two: return ViewModel(title: title, tintColor: color, image: image)
-            case .three: return ViewModel(title: title, tintColor: color, image: image)
-            case .four: return ViewModel(title: title, tintColor: color, image: image)
         }
     }
 }
